+++ console/client/src/features/verbs/VerbPage.tsx
@@ -1,15 +1,24 @@
 import { Square3Stack3DIcon } from '@heroicons/react/24/outline'
 import React from 'react'
 import { useParams } from 'react-router-dom'
+import { CodeBlock } from '../../components/CodeBlock'
 import { PageHeader } from '../../components/PageHeader'
-import { Module, Verb } from '../../protos/xyz/block/ftl/v1/console/console_pb'
+import { CallEvent, Module, Verb } from '../../protos/xyz/block/ftl/v1/console/console_pb'
 import { modulesContext } from '../../providers/modules-provider'
+import { getCalls } from '../../services/console.service'
+import { CallList } from '../calls/CallList'
+import { buildVerbSchema } from './verb.utils'
 
 export const VerbPage = () => {
   const { moduleName, verbName } = useParams()
   const modules = React.useContext(modulesContext)
   const [module, setModule] = React.useState<Module | undefined>()
   const [verb, setVerb] = React.useState<Verb | undefined>()
+  const [calls, setCalls] = React.useState<CallEvent[] | undefined>()
+
+  const callData =
+    module?.data.filter((data) => [verb?.verb?.request?.name, verb?.verb?.response?.name].includes(data.data?.name)) ??
+    []
 
   React.useEffect(() => {
     if (modules) {
@@ -20,18 +29,44 @@ export const VerbPage = () => {
     }
   }, [modules, moduleName])
 
+  React.useEffect(() => {
+    if (!module) return
+
+    const fetchCalls = async () => {
+      const calls = await getCalls(module.name, verb?.verb?.name)
+      setCalls(calls)
+    }
+    fetchCalls()
+  }, [module])
+
   return (
     <>
-      <PageHeader
-        icon={<Square3Stack3DIcon />}
-        title={verb?.verb?.name || ''}
-        breadcrumbs={[
-          { label: 'Modules', link: '/modules' },
-          { label: module?.name || '', link: `/modules/${module?.name}` },
-        ]}
-      />
-      <div className='m-4'>
-        <h1>Verb: {verb?.verb?.name}</h1>
+      <div className='flex flex-col h-full'>
+        <div className='flex-1 flex flex-col h-full'>
+          <PageHeader
+            icon={<Square3Stack3DIcon />}
+            title={verb?.verb?.name || ''}
+            breadcrumbs={[
+              { label: 'Modules', link: '/modules' },
+              { label: module?.name || '', link: `/modules/${module?.name}` },
+            ]}
+          />
+          <div className='flex-1 h-1/2 overflow-y-auto mx-4 mt-4'>
+            {verb?.verb?.request?.toJsonString() && (
+              <CodeBlock
+                code={buildVerbSchema(
+                  verb?.schema,
+                  callData.map((d) => d.schema),
+                )}
+                language='json'
+              />
+            )}
+          </div>
+
+          <div className='flex-1 h-1/2 m-4'>
+            <CallList calls={calls} />
+          </div>
+        </div>
       </div>
     </>
   )
